Counter update and storage per session

+++ src/client.cpp
@@ -7,4 +7,4 @@ std::string &Client::getPrivateKey() { return this->m_private_key; }
 
 std::string &Client::getIdentifier() { return this->m_identifier; }
 
-uint32_t Client::getCounter() { return this->m_counter++; }
+uint64_t Client::getCounter() { return this->m_counter++; }
